+++ src/AlfBasicAttribute.php
@@ -10,7 +10,7 @@ abstract class AlfBasicAttribute extends AlfBasicClass {
     /** @AlfAttrAutoComplete */
     #[AlfAttrAutoComplete]
     final public static function _AlfBasicAttribute($obj) : AlfBasicAttribute {
-        return AlfProgramming::_()->unused($obj, parent::_AlfBasicClass($obj));
+        return AlfProgramming::_()->unused($obj, static::_AlfBasicClass($obj));
     }
 
 }