+++ src/AlfBasicType.php
@@ -10,7 +10,7 @@ abstract class AlfBasicType extends AlfBasicClass {
     /** @AlfAttrAutoComplete */
     #[AlfAttrAutoComplete]
     public static function _AlfBasicType($obj) : AlfBasicType {
-        return AlfProgramming::_()->unused($obj, parent::_AlfBasicClass($obj));
+        return AlfProgramming::_()->unused($obj, static::_AlfBasicClass($obj));
     }
 
 }