+++ src/Attributes/AlfAttrAutoComplete.php
@@ -12,7 +12,7 @@ class AlfAttrAutoComplete extends AlfBasicAttribute {
     /** @AlfAttrAutoComplete */
     #[AlfAttrAutoComplete]
     final public static function _AlfAttrAutoComplete($obj) : AlfAttrAutoComplete {
-        return AlfProgramming::_()->unused($obj, parent::_AlfBasicAttribute($obj));
+        return AlfProgramming::_()->unused($obj, static::_AlfBasicAttribute($obj));
     }
 
 }