+++ src/Attributes/AlfAttrParameter.php
@@ -12,7 +12,7 @@ abstract class AlfAttrParameter extends AlfBasicAttribute {
     /** @AlfAttrAutoComplete */
     #[AlfAttrAutoComplete]
     final public static function _AlfAttrParameter($obj) : AlfAttrParameter {
-        return AlfProgramming::_()->unused($obj, parent::_AlfBasicAttribute($obj));
+        return AlfProgramming::_()->unused($obj, static::_AlfBasicAttribute($obj));
     }
 
 }