+++ src/Attributes/AlfAttrParameterIsInt.php
@@ -11,7 +11,7 @@ class AlfAttrParameterIsInt extends AlfAttrParameter {
     /** @AlfAttrAutoComplete */
     #[AlfAttrAutoComplete]
     final public static function _AlfAttrParameterIsInt($obj) : AlfAttrParameterIsInt {
-        return AlfProgramming::_()->unused($obj, parent::_AlfAttrParameter($obj));
+        return AlfProgramming::_()->unused($obj, static::_AlfAttrParameter($obj));
     }
 
 }